+++ src/hardis-websocket-server.ts
@@ -6,6 +6,7 @@ import { getWorkspaceRoot, stripAnsi } from "./utils";
 import { Logger } from "./logger";
 import { LwcPanelManager } from "./lwc-panel-manager";
 import { HardisStatusProvider } from "./hardis-status-provider";
+import { refreshDataWorkbenchPanel } from "./commands/showDataWorkbench";
 
 const DEFAULT_PORT = parseInt(process.env.SFDX_HARDIS_WEBSOCKET_PORT || "2702");
 let globalWss: LocalWebSocketServer | null;
@@ -344,13 +345,7 @@ export class LocalWebSocketServer {
     }
     // Request to refresh data workbench
     else if (data.event === "refreshDataWorkbench") {
-      const panelManager = LwcPanelManager.getInstance();
-      const dataWorkbenchPanel = panelManager.getPanel("s-data-workbench");
-      if (dataWorkbenchPanel) {
-        dataWorkbenchPanel.sendMessage({
-          type: "refreshWorkspaces",
-        });
-      }
+      refreshDataWorkbenchPanel();
     }
     // Request to refresh commands box
     else if (data.event === "runSfdxHardisCommand") {